introduce skeleton for table coding system
If the message structure is compatible, the DSDL compiler generates a
table which describes its structure. If enabled, an interpreter is then
used to follow the table description and encode or decode the message.
This occupies substantially fewer bytes per message as the table is much
smaller than the C function previously used. This translates to a large
reduction in program memory usage, especially if many different message
types need to be coded.

This commit sets up all the hooks to enable and disable the system and
defines the interpreter functions. However, they don't do anything as
the system currently doesn't support any types.