canard++: fix handler list constructor/destructor ordering
The handler subclass must not exist in the handler list until after the
superclass constructor runs, and must be removed before the superclass
destructor runs. Outside of this time, the handler object identity is
the superclass, so a call to the handle_message method will cause a pure
virtual error and crash the program.

Fix the issue by adding/removing the object in the handler list in the
subclass when the vtable is set up right and the handle_message can
safely be called. Also wait until the branch list is set up so that the
handle_message method can actually find the object.

Co-authored-by: Andrew Tridgell <andrew@tridgell.net>